fix: update code to move additional tests
fix: update code with new changes
