+++ api/src/lib.rs
@@ -9,4 +9,4 @@ pub use resume_token::ResumeToken;
 pub use server::{AccessInterestStore, AccessModelStore, Server};
 
 #[cfg(test)]
-mod test;
+mod tests;
